Move o.g.jvm.{internal.apigen=>tasks.api}
This commit moves all of the types directly within the
o.g.jvm.internal.apigen package to their new home in o.g.jvm.tasks.api.
Stay tuned for more exciting developments in subsequent commits.
